requirements.txt
huggingface-hub==0.27.0
SQLAlchemy==2.0.38
tenacity==9.0.0
tqdm==4.67.1
webvtt-py==0.5.1
nltk==3.9.1